fix number expression parse overflow

+++ Verilog/Expressions/Number.cs
@@ -136,7 +136,7 @@ namespace pluginVerilog.Verilog.Expressions
             if(apostropheIndex == -1)
             { // decimal
                 number.NumberType = NumberTypeEnum.Decimal;
-                number.Value = int.Parse(sb.ToString());
+                number.Value = long.Parse(sb.ToString());
                 number.Constant = true;
                 word.MoveNext();
                 return number;
